fix: 🐞 a bug due to destructure
