+++ packages/backend/server_generate/package.json
@@ -42,7 +42,6 @@
     "uuid": "^8.3.2",
     "winston": "^3.2.1",
     "ws": "^8.3.0"
-
   },
   "devDependencies": {
     "axios": "^0.21.1",
