+++ packages/frontend/.env.development
@@ -1,1 +1,1 @@
-REACT_APP_BASE_URL=http://10.0.0.38:8096
+REACT_APP_BASE_URL=http://localhost:8096
