+++ packages/frontend/src/App.jsx
@@ -4,7 +4,7 @@ import { Route, Switch } from 'react-router-dom';
 import { createTheme, ThemeProvider } from '@material-ui/core/styles';
 import LoginPage from './pages/LoginPage';
 import home from './pages/home';
-import PrivateRoute from './utils/components/PrivateRoute';
+import PrivateRoute from './components/layout/PrivateRoute/PrivateRoute';
 import { MuiPickersUtilsProvider } from '@material-ui/pickers';
 import DateFnsUtils from '@date-io/date-fns';
 import { setAuthToken } from './utils/constants';
@@ -14,7 +14,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import { LOGIN_SUCCESS } from './redux/auth/constants';
 import AuthorsNewArticle from './components/ui/author/AuthorsnewArticle/AuthorsNewArticle';
 import AllPublications from './components/ui/author/allPublications/AllPublications/AllPublications';
-import TopBar from './utils/components/TopBar';
+import TopBar from './components/layout/TopBar/TopBar';
 import GeneralContractView from './components/ui/salesman/contractViews/GeneralContractView/GeneralContractView';
 import Snackbar from './components/ui/Snackbar/Snackbar.jsx';
 import Sales from './components/ui/salesman/Sales/Sales';
@@ -22,7 +22,7 @@ import DeadArticle from './components/ui/author/DeadArticle/DeadArticle';
 import Article from './components/ui/researches/Article/Article';
 import SalesUsersScreen from './components/ui/admin/SalesUsersScreen/SalesUsersScreen.jsx';
 import AuthorsUsersScreen from './components/ui/admin/AuthorsUsersScreen/AuthorsUsersScreen.jsx';
-import FooterMember from './utils/components/FooterMember';
+import FooterMember from './components/layout/FooterMember/FooterMember';
 import MembersMain from './components/Members/MembersMain';
 import FullPublication from './components/Members/ShowPublication/FullPublication';
 import AllContracts from './components/ui/salesman/contractViews/allContracts/AllContract/AllContract';
