+++ packages/frontend/src/components/layout/PrivateRoute/PrivateRoute.view.jsx
@@ -1,0 +1,30 @@
+import React from 'react';
+import { Route, Redirect } from 'react-router-dom';
+
+//import useStyles from './PrivateRoute.style';
+
+const PrivateRouteView = (props) => {
+const Component = props.Component;
+  return  (
+		<Route
+			{...props.rest}
+			render={(rest) =>
+				props.isAuthenticated ? (
+					<Component {...rest} />
+				) : (
+					<Redirect
+						to={{
+							pathname: '/Login',
+							state: { from: rest.location },
+						}}
+					/>
+				)
+			}
+		/>
+	);
+};
+
+PrivateRouteView.displayName = 'PrivateRouteView';
+PrivateRouteView.defaultProps = {};
+
+export default React.memo(PrivateRouteView);
