--- packages/frontend/src/utils/components/PrivateRoute.js
@@ -1,25 +1,0 @@
-import { Route, Redirect } from 'react-router-dom';
-import { useSelector } from 'react-redux';
-
-const PrivateRoute = ({ component: Component, ...rest }) => {
-	const isAuthenticated = useSelector((state) => state.auth.isAuthenticated);
-	return (
-		<Route
-			{...rest}
-			render={(props) =>
-				isAuthenticated ? (
-					<Component {...props} />
-				) : (
-					<Redirect
-						to={{
-							pathname: '/Login',
-							state: { from: props.location },
-						}}
-					/>
-				)
-			}
-		/>
-	);
-};
-
-export default PrivateRoute;
