--- packages/frontend/src/utils/components/Select.js
@@ -1,11 +1,0 @@
-import Select from 'react-select';
-
-const options = [
-	{ value: 'chocolate', label: 'Chocolate' },
-	{ value: 'strawberry', label: 'Strawberry' },
-	{ value: 'vanilla', label: 'Vanilla' },
-];
-
-const MySelect = () => <Select options={options} />;
-
-export default MySelect;
